Move createNumberMask to utils
